+++ lib/routes/parse.go
@@ -1,71 +1,55 @@
 package routes
 
 import (
-	"net/http"
 	"strings"
 )
 
-func EnableRouter(routes *Routes) {
-	http.HandleFunc("/", func(res http.ResponseWriter, req *http.Request) {
-		path := strings.TrimRight(strings.TrimLeft(req.URL.Path, "/"), "/")
-		parts := strings.Split(path, "/")
+func parse(routes *Routes, path string) (*Route, RouteParams) {
+	path = strings.TrimRight(strings.TrimLeft(path, "/"), "/")
+	parts := strings.Split(path, "/")
 
-		// Parse the Route
-		var route *Route = nil
-		var node Routes = *routes
-		params := make(RouteParams)
+	// Parse the Route
+	var route *Route = nil
+	var node Routes = *routes
+	params := make(RouteParams)
 
-		for i, part := range parts {
+	for i, part := range parts {
 
-			// If the last part try to get the route
-			if i == len(parts)-1 {
-				// Check if exist in node
-				if _, ok := node[part]; ok {
-					route = node[part]
-					continue
-				}
-
-				// Check if in node routes exist a special route
-				if _, ok := node["@"]; ok {
-					route = node["@"]
-					params[route.Param] = part
-					continue
-				}
-			}
-
-			// Check if exist in node routes
+		// If the last part try to get the route
+		if i == len(parts)-1 {
+			// Check if exist in node
 			if _, ok := node[part]; ok {
-				node = node[part].Routes
+				route = node[part]
 				continue
 			}
 
 			// Check if in node routes exist a special route
 			if _, ok := node["@"]; ok {
-				r := node["@"]
-
-				params[r.Param] = part
-				node = r.Routes
+				route = node["@"]
+				params[route.Param] = part
 				continue
 			}
+		}
 
-			// Nullifier route
-			route = nil
+		// Check if exist in node routes
+		if _, ok := node[part]; ok {
+			node = node[part].Routes
 			continue
 		}
 
-		// End of Parser
+		// Check if in node routes exist a special route
+		if _, ok := node["@"]; ok {
+			r := node["@"]
 
-		if route == nil || route.Handler == nil {
-			http.NotFound(res, req)
-			return
+			params[r.Param] = part
+			node = r.Routes
+			continue
 		}
 
-		context := DWorkContext{
-			Params:   params,
-			Response: res,
-			Request:  req,
-		}
+		// Nullifier route
+		route = nil
+		continue
+	}
 
-		route.Handler(context)
-	})
+	return route, params
 }